Resolve S12P31S104-71 "Feature/아카이브 채팅 추가 구현"

See merge request s12-final/S12P31S104!25

+++ dbdeep-BE/src/main/java/kr/dbdeep/dbdeep_BE/domain/chat/application/ChatShareService.java
@@ -78,7 +78,7 @@ public class ChatShareService {
                         .id(UUID.randomUUID().toString())
                         .chatRoomId(newChatRoomId)
                         .content(m.getContent())
-                        .memberId(memberId.longValue())
+                        .memberId((int) memberId.longValue())
                         .senderType(m.getSenderType())
                         .timestamp(m.getTimestamp())
                         .build())
